requirements.txt
flask==3.0.3
flask-cors==4.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
anthropic>=0.49.0
sentence-transformers==3.0.1
pgvector==0.3.2
python-dotenv==1.0.1
gunicorn==22.0.0
numpy==1.26.4
requests==2.32.3
